Fixing 64 bit issues with large data sets. Fixing compiler warning for casting from 64 bit int to 32 bit integer

+++ Source/DREAM3DLib/ProcessingFilters/MinNeighbors.cpp
@@ -334,7 +334,7 @@ void MinNeighbors::assign_badpoints()
       }
     }
     QList<QString> voxelArrayNames = m->getCellArrayNameList();
-    for (size_t j = 0; j < totalPoints; j++)
+    for (int64_t j = 0; j < totalPoints; j++)
     {
       grainname = m_GrainIds[j];
       neighbor = m_Neighbors[j];
